requirements.txt
discord.py
pycoingecko
requests